// src/domain/enums/GithubListEnums.ts
/**
 * @enum {GithubListEnums} GithubListEnums - Definição de tipo de unidade de armazenamento
 *
 * @arg {repository} REPOSITORY - Repository
 * @arg {gist} GIST - Gist
 * @arg {follower} FOLLOWER - Follower
 * @arg {following} FOLLOWING - Following
 */

export enum GithubListEnums {
  REPOSITORY = 'repository',
  GIST = 'gist',
  FOLLOWER = 'follower',
  FOLLOWING = 'following',
}